refactor(splicetcp): split tcp bridge
